[SC-350] Moved email properties and some fixes
