+++ api/src/main/java/com/example/api/service/util/email/EmailProperties.java
@@ -1,0 +1,15 @@
+package com.example.api.service.util.email;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ConfigurationProperties(prefix = "email")
+@Getter
+@Setter
+public class EmailProperties {
+    private String apiKey;
+    private String apiSecretKey;
+}
