+++ api/src/main/resources/application.properties
@@ -10,3 +10,5 @@ spring.jpa.properties.hibernate.format_sql=true
 spring.mvc.pathmatch.matching-strategy=ant-path-matcher
 spring.servlet.multipart.max-file-size=5MB
 spring.servlet.multipart.max-request-size=5MB
+email.api-key=
+email.api-secret-key=
